requirements.txt
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
